Old test submission.

+++ old_test_submission.py
@@ -6,13 +6,9 @@
 # * Resources might vary from your local machine
 
 from typing import NamedTuple, List
-from dataclasses import dataclass
-from multiprocessing import Process, Queue
-import os
-import queue
-
 import numpy as np
 import pandas as pd
+import os
 
 from submission_config import SubmissionConfig, TestEvaluationConfig
 
@@ -25,12 +21,16 @@ import parse_ttyrec
 class RolloutResults(NamedTuple):
     ascensions: int
     scores: List[int]
-    log_paths: str
+    median_score: int
+    mean_score: int
+    log_path: str
 
-def evaluate(id, num_episodes=TestEvaluationConfig.NUM_EPISODES, runner_queue=None):
+def evaluate():
     env_make_fn = SubmissionConfig.MAKE_ENV_FN
     num_envs = SubmissionConfig.NUM_ENVIRONMENTS
     Agent = SubmissionConfig.AGENT
+
+    num_episodes = TestEvaluationConfig.NUM_EPISODES
 
     batched_env = BatchedEnv(env_make_fn=env_make_fn, num_envs=num_envs)
 
@@ -38,72 +38,26 @@ def evaluate(id, num_episodes=TestEvaluationConfig.NUM_EPISODES, runner_queue=No
 
     ascensions, scores = run_batched_rollout(num_episodes, batched_env, agent)
 
-    log_paths = []
+    log_path = None
     if environment.env.log_runs:
-        log_paths.append(batched_env.envs[0].savedir)
+        log_path = batched_env.envs[0].savedir
 
     results = RolloutResults(
         ascensions=ascensions,
         scores=scores,
-        log_paths=log_paths,
+        median_score=np.median(scores),
+        mean_score=np.mean(scores),
+        log_path=log_path,
     )
-
-    if runner_queue:
-        runner_queue.put(results)
 
     return results
 
-@dataclass
-class Runner:
-    id: int
-    process: Process
-    queue: Queue
-    done: bool = False
-
-def merge_results(results_1: RolloutResults, results_2: RolloutResults):
-    return RolloutResults(
-        ascensions=results_1.ascensions + results_2.ascensions,
-        scores=results_1.scores + results_2.scores,
-        log_paths=results_1.log_paths + results_2.log_paths,
-    )
 
 if __name__ == "__main__":
-    overall_results = RolloutResults(
-        ascensions=0,
-        scores=[],
-        log_paths=[],
-    )
-    runners : List[Runner] = []
-    episodes_per_runner = TestEvaluationConfig.NUM_EPISODES // environment.env.num_runners + 1
-    for i in range(0, environment.env.num_runners):
-        runner_queue = Queue()
-        runner = Runner(
-            id=i,
-            process=Process(target=evaluate, args=(i, episodes_per_runner, runner_queue)),
-            queue=runner_queue,
-        )
-        runners.append(runner)
+    results = evaluate()
 
-    for runner in runners:
-        runner.process.start()
-
-    done_runners = 0
-
-    while done_runners < environment.env.num_runners:
-        for runner in runners:
-            if runner.done:
-                continue
-            try:
-                new_results = runner.queue.get(timeout=1)
-                overall_results = merge_results(overall_results, new_results)
-            except queue.Empty:
-                pass
-            runner.process.join(timeout=1)
-            if runner.process.exitcode is not None:
-                runner.done = True
-                done_runners += 1
-
-    for path in overall_results.log_paths:
+    if results.log_path is not None:
+        path = results.log_path
         files = [os.path.join(path,f) for f in os.listdir(path) if os.path.isfile(os.path.join(path,f)) and f.endswith('.ttyrec.bz2')]
         for f in files:
             if f.endswith('{}.ttyrec.bz2'.format(environment.env.num_episodes)): # rm this junk file
